// src/models/medicine.model.ts
import { REGULATORY_CATEGORY, REGISTRATION_STATUS } from "./enums";

export interface Medicine {
  PRODUCT_TYPE: string;
  PRODUCT_NAME: string;
  PROCESS_FINALIZATION_DATE: string;
  REGULATORY_CATEGORY: REGULATORY_CATEGORY;
  PRODUCT_REGISTRATION_NUMBER: string;
  REGISTRATION_EXPIRATION_DATE: string;
  PROCESS_NUMBER: string;
  THERAPEUTIC_CLASS: string;
  REGISTRATION_HOLDER_COMPANY: string;
  REGISTRATION_STATUS: REGISTRATION_STATUS;
  ACTIVE_INGREDIENT: string;
}